reorganize selector util helpers
